Extract event calendar components
