+++ src/Pages/Events/Calendar/calendarConstants.js
@@ -1,0 +1,9 @@
+export const DAYS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+
+export const MONTHS = [
+  'January', 'February', 'March', 'April', 'May', 'June',
+  'July', 'August', 'September', 'October', 'November', 'December',
+];
+
+const currentYear = new Date().getFullYear();
+export const YEAR_RANGE = Array.from({ length: 10 }, (_, i) => currentYear + i);
